refactor: move UI and MainActivity to presentation package
